+++ parler/cache.py
@@ -41,9 +41,12 @@ def get_object_cache_keys(instance):
         return []
 
     keys = []
+    tr_models = instance._parler_meta.get_all_models()
+
     # TODO: performs a query to fetch the language codes. Store that in memcached too.
     for language in instance.get_available_languages():
-        keys.append(get_translation_cache_key(instance._parler_meta.translations_model, instance.pk, language))
+        for tr_model in tr_models:
+            keys.append(get_translation_cache_key(tr_model, instance.pk, language))
 
     return keys
 
@@ -57,18 +60,21 @@ def get_translation_cache_key(translated_model, master_id, language_code):
     return 'parler.{0}.{1}.{2}.{3}'.format(translated_model._meta.app_label, translated_model.__name__, long(master_id), language_code)
 
 
-def get_cached_translation(instance, language_code=None, use_fallback=False):
+def get_cached_translation(instance, language_code=None, related_name=None, use_fallback=False):
     """
     Fetch an cached translation.
+
+    .. versionadded 1.2 Added the ``related_name`` parameter.
     """
     if language_code is None:
         language_code = instance.get_current_language()
 
-    values = _get_cached_values(instance, language_code, use_fallback)
+    translated_model = instance._parler_meta.get_model_by_related_name(related_name)
+    values = _get_cached_values(instance, translated_model, language_code, use_fallback)
     if not values:
         return None
 
-    translation = instance._parler_meta.translations_model(**values)
+    translation = translated_model(**values)
     translation._state.adding = False
     return translation
 
@@ -85,7 +91,8 @@ def get_cached_translated_field(instance, field_name, language_code=None, use_fa
     if len(field_name) <= 5 and len(language_code) > 5:
         raise RuntimeError("Unexpected language code, did you swap field_name, language_code?")
 
-    values = _get_cached_values(instance, language_code, use_fallback)
+    translated_model = instance._parler_meta.get_model_by_field(field_name)
+    values = _get_cached_values(instance, translated_model, language_code, use_fallback)
     if not values:
         return None
 
@@ -93,14 +100,14 @@ def get_cached_translated_field(instance, field_name, language_code=None, use_fa
     return values.get(field_name, None)
 
 
-def _get_cached_values(instance, language_code, use_fallback=False):
+def _get_cached_values(instance, translated_model, language_code, use_fallback=False):
     """
     Fetch an cached field.
     """
     if not appsettings.PARLER_ENABLE_CACHING or not instance.pk or instance._state.adding:
         return None
 
-    key = get_translation_cache_key(instance._parler_meta.translations_model, instance.pk, language_code)
+    key = get_translation_cache_key(translated_model, instance.pk, language_code)
     values = cache.get(key)
     if not values:
         return None
@@ -108,13 +115,13 @@ def _get_cached_values(instance, language_code, use_fallback=False):
     # Check for a stored fallback marker
     if values.get('__FALLBACK__', False):
         # Internal trick, already set the fallback marker, so no query will be performed.
-        instance._translations_cache[language_code] = MISSING
+        instance._translations_cache[translated_model][language_code] = MISSING
 
         # Allow to return the fallback language instead.
         if use_fallback:
             lang_dict = get_language_settings(language_code)
             if lang_dict['fallback'] != language_code:
-                return _get_cached_values(instance, lang_dict['fallback'], use_fallback=False)
+                return _get_cached_values(instance, translated_model, lang_dict['fallback'], use_fallback=False)
         return None
 
     values['master'] = instance
@@ -144,20 +151,20 @@ def _cache_translation(translation, timeout=DEFAULT_TIMEOUT):
 
 
 
-def _cache_translation_needs_fallback(instance, language_code, timeout=DEFAULT_TIMEOUT):
+def _cache_translation_needs_fallback(instance, language_code, related_name, timeout=DEFAULT_TIMEOUT):
     """
     Store the fact that a translation doesn't exist, and the fallback should be used.
     """
     if not appsettings.PARLER_ENABLE_CACHING or not instance.pk or instance._state.adding:
         return
 
-    key = get_translation_cache_key(instance._parler_meta.translations_model, instance.pk, language_code)
+    tr_model = instance._parler_meta.get_model_by_related_name(related_name)
+    key = get_translation_cache_key(tr_model, instance.pk, language_code)
     cache.set(key, {'__FALLBACK__': True}, timeout=timeout)
 
 
 def _delete_cached_translations(shared_model):
-    for key in get_object_cache_keys(shared_model):
-        cache.delete(key)
+    cache.delete_many(get_object_cache_keys(shared_model))
 
 
 def _delete_cached_translation(translation):