+++ parler/tests/__init__.py
@@ -1,6 +1,7 @@
 from .admin import *
 from .model_construction import *
 from .model_attributes import *
+from .model_inheritance import *
 from .forms import *
 from .urls import *
 from .query_count import *